requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
Flask-Login>=0.6,<1.0
Flask-WTF>=1.2,<2.0
Flask-Bcrypt>=1.0,<2.0
Flask-Caching>=2.1,<3.0
Flask-APScheduler>=1.13,<2.0
Authlib>=1.3,<2.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=21.0,<22.0
psycopg2-binary>=2.9,<3.0
rapidfuzz>=3.0,<4.0
WTForms>=3.1,<4.0
email-validator>=2.0,<3.0
